add jinja syntax specification
