+++ Cargo.toml
@@ -47,7 +47,7 @@ serde_with = "3.0.0"
 url = "2.4.0"
 tracing = "0.1.37"
 clap = { version = "4.3.3", features = ["derive", "env", "cargo"] }
-minijinja = { version = "1.0.0-alpha.2", features = ["unstable_machinery"] }
+minijinja = { version = "1.0.0-alpha.2", features = ["unstable_machinery", "custom_syntax"] }
 tracing-subscriber = { version = "0.3.17", features = ["env-filter"] }
 num_cpus = "1.15.0"
 goblin = "0.7.1"
