+++ src/render/mod.rs
@@ -1,4 +1,5 @@
 pub mod dependency_list;
+pub mod jinja;
 pub mod pin;
 pub mod recipe;
 pub mod resolved_dependencies;
